+++ apps/platform/src/app/(main)/(project)/[workspace]/[project]/@variable/page.tsx
@@ -2,8 +2,8 @@
 
 import { useEffect, useState } from 'react'
 import { useAtom, useAtomValue, useSetAtom } from 'jotai'
-import { VariableSVG } from '@public/svg/dashboard'
 import { toast } from 'sonner'
+import { VariableSVG } from '@public/svg/dashboard'
 import {
   createVariableOpenAtom,
   selectedProjectAtom,
@@ -20,8 +20,7 @@ import { Button } from '@/components/ui/button'
 import { Accordion } from '@/components/ui/accordion'
 import { useHttp } from '@/hooks/use-http'
 import VariableLoader from '@/components/dashboard/variable/variableLoader'
-
-const ITEMS_PER_PAGE = 10
+import { VARIABLES_PAGE_SIZE } from '@/lib/constants'
 
 function VariablePage(): React.JSX.Element {
   const [page, setPage] = useState(0)
@@ -38,7 +37,7 @@ function VariablePage(): React.JSX.Element {
     ControllerInstance.getInstance().variableController.getAllVariablesOfProject({
         projectSlug: selectedProject!.slug,
         page,
-        limit: ITEMS_PER_PAGE,
+        limit: VARIABLES_PAGE_SIZE,
     }, {})
   )