+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [2.13.0-stage.4](https://github.com/keyshade-xyz/keyshade/compare/v2.13.0-stage.3...v2.13.0-stage.4) (2025-03-05)
+
+### 🚀 Features
+
+* **platform:** Create and Delete API Keys ([#726](https://github.com/keyshade-xyz/keyshade/issues/726)) ([3b1cfe7](https://github.com/keyshade-xyz/keyshade/commit/3b1cfe7ed67e7486df313a1177c403de174e359b))
+
+## [2.13.0-stage.3](https://github.com/keyshade-xyz/keyshade/compare/v2.13.0-stage.2...v2.13.0-stage.3) (2025-03-03)
+
+### 🚀 Features
+
+* **platform:** Add pagination & loading more functionality for [secure]s page ([#809](https://github.com/keyshade-xyz/keyshade/issues/809)) ([ab58f31](https://github.com/keyshade-xyz/keyshade/commit/ab58f314497d0084e5f0058888c44bb0f272a3a6))
+
 ## [2.13.0-stage.2](https://github.com/keyshade-xyz/keyshade/compare/v2.13.0-stage.1...v2.13.0-stage.2) (2025-03-01)
 
 ### 🚀 Features